stepartist textbox more resembles SL 3.95, move PeakNPS
-Replaces the stepartist quad with an ActorMultiVertex similar to the "COUNTS AS X ROUNDS" bubble. This asset is also longer, spanning the entire width of DensityGraph (length adjusted smaller to fit in 4:3).

-Used the same animation of the stepartist bubble flying in/out as in SL3.95 .

-Reorder the loading of StepsDisplayList and and PerPlayer so that the the stepartistbubble loads on top of the StepsDisplayList.

-Append "Peak NPS" readout to the stream breakdown to allow for more room for the Stepartist field. (PeakNPS no longer appears to be "floating" above the NPS graph.)

I originally wanted to stretch the stepartist bubble to be long enough to reach PeakNPS, but I couldn't figure out the layering conflicts whereby PeakNPS and the NPSgraph are in the same ActorFrame, yet PeakNPS would have to draw higher than the Stepartist bubble while the bubble also draws on top of the NPSgraph. Since draworder doesn't work to fix the layering conflict, the only way to fix this would be to completely merge the two.