[OF] New GetTotalScoresWithGrade function
Courtesy of @concubidated, a new function was added to Outfox to tally all player scores of a particular rank. Previously, a loop had to be used, but the engine was very slow to calculate all the scores; the previous method added about 20 seconds of loading to ScreenSelectMusic and 5 MINUTES to exiting the game from ScreenTitleMenu.

As of now, GetTotalScoresWithGrade tallies all scores found in Stats.xml (or ECFA_Stats.xml) that are found in the player profile regardless of whether that particular song is found in Songman(), across all difficulties, across all playstyles. This means that scores will be tallied whether or not the particular song is loaded, meaning that this function is suitable for USB profiles in addition to local profiles. As of now, this function relies on player profiles being saved to report new score tallies.

This function may change in the future.

Co-Authored-By: Tyler Brekke <concubidated@users.noreply.github.com>